refactor(schema): vitals統合・monetary integer化・構造修正
## スキーマ変更
- vitals + vital_records を統合して vital_records 1テーブルに統一
  - pet_id を直接保持（医療記録・入院記録への nullable FK + CHECK 制約）
  - body_weight_unit ENUM 追加（Kg/Lbs）
- checkups/exams/daily_records に clinic_id 追加（マルチテナント完全対応）
- 全monetary カラム numeric→integer（JPY単位不要）
- exam_type_items に unit, updated_at 追加
- trimming_records の weight 削除、bw/bt を numeric 型に変更
- staff_note_records.time を text→time 型に修正
- hospitalizations, reservation_appointments に日付整合 CHECK 制約追加
- treatments/estimates/billing_items に quantity>0 CHECK 制約追加
- billings に金額 >= 0 CHECK 制約追加
- vaccinations/trimming_records/checkups/exams の pet_id ON DELETE SET NULL → RESTRICT
- 35+ FK インデックス追加
- 14 UNIQUE インデックス（master tables の clinic_id + name）

## Goモデル変更
- vital.go: Vital→VitalRecord に完全書き換え
- hospitalization.go: DailyRecord に ClinicID 追加、TreatmentPlan/CarePlanItem monetary int64 化
- checkup_record.go: Checkup に ClinicID 追加
- examination_record.go: Examination に ClinicID、ExaminationTypeItem に Unit/UpdatedAt 追加
- trimming.go: Weight 削除、BW/BT を *float64 に変更
- treatment.go/accounting.go/estimate.go: monetary フィールドを int64 に変更
- 全マスタモデル: Price *float64 → *int64
- handler/*_request.go, *_response.go: 型変更に追従
- service/*.go: monetary/DTO 型変更に追従

## codegen
- models.ts 再生成（VitalRecord 反映・int64 型変更反映）

+++ backend/internal/handler/accounting_response.go
@@ -11,7 +11,7 @@ type billingItemResponse struct {
 	BillingID             uint64    `json:"billing_id"`
 	Category              string    `json:"category"`
 	Name                  string    `json:"name"`
-	UnitPrice             float64   `json:"unit_price"`
+	UnitPrice             int64     `json:"unit_price"`
 	Quantity              float64   `json:"quantity"`
 	TaxRate               float64   `json:"tax_rate"`
 	IsInsuranceApplicable bool      `json:"is_insurance_applicable"`
@@ -23,16 +23,16 @@ type billingItemResponse struct {
 type paymentResponse struct {
 	ID              uint64    `json:"id"`
 	BillingID       uint64    `json:"billing_id"`
-	Subtotal        float64   `json:"subtotal"`
-	TaxTotal        float64   `json:"tax_total"`
-	TotalAmount     float64   `json:"total_amount"`
+	Subtotal        int64     `json:"subtotal"`
+	TaxTotal        int64     `json:"tax_total"`
+	TotalAmount     int64     `json:"total_amount"`
 	InsuranceName   string    `json:"insurance_name"`
 	InsuranceRatio  float64   `json:"insurance_ratio"`
-	InsuranceAmount float64   `json:"insurance_amount"`
-	DiscountAmount  float64   `json:"discount_amount"`
-	BillingAmount   float64   `json:"billing_amount"`
-	ReceivedAmount  float64   `json:"received_amount"`
-	ChangeAmount    float64   `json:"change_amount"`
+	InsuranceAmount int64     `json:"insurance_amount"`
+	DiscountAmount  int64     `json:"discount_amount"`
+	BillingAmount   int64     `json:"billing_amount"`
+	ReceivedAmount  int64     `json:"received_amount"`
+	ChangeAmount    int64     `json:"change_amount"`
 	Method          string    `json:"method"`
 	CreatedAt       time.Time `json:"created_at"`
 	UpdatedAt       time.Time `json:"updated_at"`
