+++ backend/internal/handler/exam_type_response.go
@@ -20,7 +20,7 @@ type examTypeResponse struct {
 	ID          uint64                 `json:"id"`
 	ClinicID    uint64                 `json:"clinic_id"`
 	Name        string                 `json:"name"`
-	Price       *float64               `json:"price,omitempty"`
+	Price       *int64                 `json:"price,omitempty"`
 	IsActive    bool                   `json:"is_active"`
 	Description string                 `json:"description"`
 	SortOrder   int                    `json:"sort_order"`
