+++ backend/internal/handler/hospitalization_plan_request.go
@@ -3,7 +3,7 @@ package handler
 // createHospitalizationPlanRequest は入院プラン作成リクエスト。
 type createHospitalizationPlanRequest struct {
 	Name        string   `json:"name"         binding:"required"`
-	Price       *float64 `json:"price"`
+	Price       *int64 `json:"price"`
 	IsActive    bool     `json:"is_active"`
 	Description string   `json:"description"`
 	BodySize    string   `json:"body_size"`
@@ -14,7 +14,7 @@ type createHospitalizationPlanRequest struct {
 // updateHospitalizationPlanRequest は入院プラン更新リクエスト。
 type updateHospitalizationPlanRequest struct {
 	Name        string   `json:"name"`
-	Price       *float64 `json:"price"`
+	Price       *int64 `json:"price"`
 	IsActive    *bool    `json:"is_active"`
 	Description string   `json:"description"`
 	BodySize    string   `json:"body_size"`
