+++ backend/internal/handler/medicine_request.go
@@ -2,16 +2,16 @@ package handler
 
 // createMedicineRequest は薬剤作成のバインド struct
 type createMedicineRequest struct {
-	Name            string   `json:"name"             binding:"required"`
-	ParentID        *uint64  `json:"parent_id"`
-	Price           *float64 `json:"price"`
-	IsActive        bool     `json:"is_active"`
-	Description     string   `json:"description"`
-	DosageForm      *string  `json:"dosage_form"`
-	MedicineUnit    *string  `json:"medicine_unit"`
-	InventoryID     *uint64  `json:"inventory_id"`
-	DefaultQuantity float64  `json:"default_quantity"`
-	SortOrder       int      `json:"sort_order"`
+	Name            string  `json:"name"             binding:"required"`
+	ParentID        *uint64 `json:"parent_id"`
+	Price           *int64  `json:"price"`
+	IsActive        bool    `json:"is_active"`
+	Description     string  `json:"description"`
+	DosageForm      *string `json:"dosage_form"`
+	MedicineUnit    *string `json:"medicine_unit"`
+	InventoryID     *uint64 `json:"inventory_id"`
+	DefaultQuantity float64 `json:"default_quantity"`
+	SortOrder       int     `json:"sort_order"`
 }
 
 // reorderMedicineRequest は薬剤並び替えのバインド struct
@@ -27,7 +27,7 @@ type updateMedicineRequest struct {
 	Name            *string  `json:"name"`
 	ParentID        *uint64  `json:"parent_id"`
 	ClearParentID   bool     `json:"clear_parent_id"`
-	Price           *float64 `json:"price"`
+	Price           *int64   `json:"price"`
 	IsActive        *bool    `json:"is_active"`
 	Description     *string  `json:"description"`
 	DosageForm      *string  `json:"dosage_form"`
