+++ backend/internal/handler/procedure_request.go
@@ -2,7 +2,7 @@ package handler
 
 type createProcedureRequest struct {
 	Name        string   `json:"name"        binding:"required"`
-	Price       *float64 `json:"price"`
+	Price       *int64 `json:"price"`
 	IsActive    bool     `json:"is_active"`
 	Description string   `json:"description"`
 	Duration    *int     `json:"duration"`
@@ -13,7 +13,7 @@ type createProcedureRequest struct {
 
 type updateProcedureRequest struct {
 	Name          string   `json:"name"`
-	Price         *float64 `json:"price"`
+	Price         *int64 `json:"price"`
 	IsActive      *bool    `json:"is_active"`
 	Description   string   `json:"description"`
 	Duration      *int     `json:"duration"`
