+++ backend/internal/handler/treatment_plan_request.go
@@ -4,11 +4,11 @@ type createTreatmentPlanRequest struct {
 	TreatmentContent string  `json:"treatment_content" binding:"required"`
 	Memo             string  `json:"memo"`
 	Insurance        bool    `json:"insurance"`
-	UnitPrice        float64 `json:"unit_price"`
+	UnitPrice        int64   `json:"unit_price"`
 	Quantity         float64 `json:"quantity"`
 	DiscountRate     float64 `json:"discount_rate"`
-	DiscountAmount   float64 `json:"discount_amount"`
-	Subtotal         float64 `json:"subtotal"`
+	DiscountAmount   int64   `json:"discount_amount"`
+	Subtotal         int64   `json:"subtotal"`
 	SortOrder        int     `json:"sort_order"`
 }
 
@@ -16,10 +16,10 @@ type updateTreatmentPlanRequest struct {
 	TreatmentContent *string  `json:"treatment_content"`
 	Memo             *string  `json:"memo"`
 	Insurance        *bool    `json:"insurance"`
-	UnitPrice        *float64 `json:"unit_price"`
+	UnitPrice        *int64   `json:"unit_price"`
 	Quantity         *float64 `json:"quantity"`
 	DiscountRate     *float64 `json:"discount_rate"`
-	DiscountAmount   *float64 `json:"discount_amount"`
-	Subtotal         *float64 `json:"subtotal"`
+	DiscountAmount   *int64   `json:"discount_amount"`
+	Subtotal         *int64   `json:"subtotal"`
 	SortOrder        *int     `json:"sort_order"`
 }
