+++ backend/internal/handler/treatment_plan_response.go
@@ -14,11 +14,11 @@ type treatmentPlanResponse struct {
 	TreatmentContent  string    `json:"treatment_content"`
 	Memo              string    `json:"memo"`
 	Insurance         bool      `json:"insurance"`
-	UnitPrice         float64   `json:"unit_price"`
+	UnitPrice         int64     `json:"unit_price"`
 	Quantity          float64   `json:"quantity"`
 	DiscountRate      float64   `json:"discount_rate"`
-	DiscountAmount    float64   `json:"discount_amount"`
-	Subtotal          float64   `json:"subtotal"`
+	DiscountAmount    int64     `json:"discount_amount"`
+	Subtotal          int64     `json:"subtotal"`
 	SortOrder         int       `json:"sort_order"`
 	CreatedAt         time.Time `json:"created_at"`
 	UpdatedAt         time.Time `json:"updated_at"`
