+++ backend/internal/handler/treatment_request.go
@@ -6,7 +6,7 @@ type createTreatmentRequest struct {
 	ProcedureID    *uint64 `json:"procedure_id"`
 	MedicineID     *uint64 `json:"medicine_id"`
 	InventoryID    *uint64 `json:"inventory_id"`
-	UnitPrice      float64 `json:"unit_price"`
+	UnitPrice      int64   `json:"unit_price"`
 	Quantity       float64 `json:"quantity"`
 	Selected       bool    `json:"selected"`
 	Status         string  `json:"status"`
@@ -14,7 +14,7 @@ type createTreatmentRequest struct {
 	Memo           string  `json:"memo"`
 	Insurance      bool    `json:"insurance"`
 	DiscountRate   float64 `json:"discount_rate"`
-	DiscountAmount float64 `json:"discount_amount"`
+	DiscountAmount int64   `json:"discount_amount"`
 	SortOrder      int     `json:"sort_order"`
 }
 
@@ -24,7 +24,7 @@ type updateTreatmentRequest struct {
 	ProcedureID    *uint64  `json:"procedure_id"`
 	MedicineID     *uint64  `json:"medicine_id"`
 	InventoryID    *uint64  `json:"inventory_id"`
-	UnitPrice      *float64 `json:"unit_price"`
+	UnitPrice      *int64   `json:"unit_price"`
 	Quantity       *float64 `json:"quantity"`
 	Selected       *bool    `json:"selected"`
 	Status         *string  `json:"status"`
@@ -32,7 +32,7 @@ type updateTreatmentRequest struct {
 	Memo           *string  `json:"memo"`
 	Insurance      *bool    `json:"insurance"`
 	DiscountRate   *float64 `json:"discount_rate"`
-	DiscountAmount *float64 `json:"discount_amount"`
+	DiscountAmount *int64   `json:"discount_amount"`
 	SortOrder      *int     `json:"sort_order"`
 }
